+++ services/workflows-service/src/workflow/hook-callback-handler.service.ts
@@ -158,7 +158,7 @@ export class HookCallbackHandlerService {
 
     const { context } = workflowRuntime;
     const { reportData: unvalidatedReportData, base64Pdf, reportId, reportType } = data;
-    const reportData = ReportWithRiskScoreSchema;
+    const reportData = ReportWithRiskScoreSchema.parse(unvalidatedReportData);
 
     const { documents, pdfReportBallerineFileId } =
       await this.__peristPDFReportDocumentWithWorkflowDocuments({
@@ -168,8 +168,7 @@ export class HookCallbackHandlerService {
         base64PDFString: base64Pdf as string,
       });
 
-    const reportRiskScore =
-      ReportWithRiskScoreSchema.parse(unvalidatedReportData).summary.riskScore;
+    const reportRiskScore = reportData?.summary?.riskScore;
 
     const business = await this.businessService.getByCorrelationId(context.entity.id, [
       currentProjectId,