+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -69,6 +69,7 @@ import {
 } from '@nestjs/common';
 import {
   ApprovalState,
+  EndUser,
   Prisma,
   PrismaClient,
   UiDefinitionContext,
@@ -95,6 +96,7 @@ import { addPropertiesSchemaToDocument } from './utils/add-properties-schema-to-
 import { WorkflowEventEmitterService } from './workflow-event-emitter.service';
 import { WorkflowRuntimeDataRepository } from './workflow-runtime-data.repository';
 import dayjs from 'dayjs';
+import { entitiesUpdate } from './utils/entities-update';
 
 type TEntityId = string;
 
@@ -1814,6 +1816,8 @@ export class WorkflowService {
   ): Promise<TEntityId | null> {
     if (entity.ballerineEntityId) {
       return entity.ballerineEntityId as TEntityId;
+    } else if ('data' in entity && isObject(entity.data) && entity.data.ballerineEntityId) {
+      return entity.data.ballerineEntityId as TEntityId;
     } else if (!entity.id) {
       return null;
     } else {
@@ -1876,12 +1880,14 @@ export class WorkflowService {
 
     return await beginTransactionIfNotExist(async transaction => {
       this.logger.log('Workflow event received', { id, type });
+
       const workflowRuntimeData = await this.workflowRuntimeDataRepository.findByIdAndLock(
         id,
         {},
         projectIds,
         transaction,
       );
+
       const workflowDefinition = await this.workflowDefinitionRepository.findById(
         workflowRuntimeData.workflowDefinitionId,
         {},
@@ -1930,6 +1936,31 @@ export class WorkflowService {
         },
       });
 
+      service.subscribe('ENTITIES_UPDATE', async ({ payload }) => {
+        if (
+          !payload ||
+          !payload.ubos ||
+          !payload.directors ||
+          !Array.isArray(payload.ubos) ||
+          !Array.isArray(payload.directors)
+        ) {
+          return;
+        }
+
+        const typedPayload = payload as {
+          ubos: Array<Partial<EndUser>>;
+          directors: Array<Partial<EndUser>>;
+        };
+
+        await entitiesUpdate({
+          endUserService: this.endUserService,
+          projectId: currentProjectId,
+          businessId: workflowRuntimeData.businessId,
+          sendEvent: service.sendEvent,
+          payload: typedPayload,
+        });
+      });
+
       if (!service.getSnapshot().nextEvents.includes(type)) {
         throw new BadRequestException(
           `Event ${type} does not exist for workflow ${workflowDefinition.id}'s state: ${workflowRuntimeData.state}`,
@@ -1943,9 +1974,10 @@ export class WorkflowService {
 
       const snapshot = service.getSnapshot();
       const currentState = snapshot.value;
-      const context = snapshot.machine.context;
+      const context = snapshot.machine?.context;
       // TODO: Refactor to use snapshot.done instead
-      const isFinal = snapshot.machine.states[currentState].type === 'final';
+      // @ts-ignore
+      const isFinal = snapshot.machine?.states[currentState].type === 'final';
       const entityType = aliasIndividualAsEndUser(context?.entity?.type);
       const entityId = workflowRuntimeData[`${entityType}Id`];
 
@@ -1959,6 +1991,7 @@ export class WorkflowService {
         workflowRuntimeData.id,
         {
           context,
+          // @ts-ignore
           state: currentState,
           tags: Array.from(snapshot.tags) as unknown as WorkflowDefinitionUpdateInput['tags'],
           status: isFinal ? 'completed' : workflowRuntimeData.status,
@@ -1975,6 +2008,7 @@ export class WorkflowService {
           projectIds,
           currentProjectId,
           transaction,
+          // @ts-ignore
           currentState,
         );
       }