+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ballerine/backoffice-v2
 
+## 0.7.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.6.11
+  - @ballerine/workflow-node-sdk@0.6.11
+
+## 0.7.9
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.10
+- @ballerine/workflow-node-sdk@0.6.10
+
 ## 0.7.8
 
 ### Patch Changes