+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,18 @@
 # kyb-app
 
+## 0.3.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.6.11
+
+## 0.3.10
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.10
+
 ## 0.3.9
 
 ### Patch Changes