+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ballerine/headless-example
 
+## 0.3.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.6.11
+
+## 0.3.9
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.10
+
 ## 0.3.8
 
 ### Patch Changes