+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/workflow-core
 
+## 0.6.11
+
+### Patch Changes
+
+- Version bump
+
+## 0.6.10
+
+### Patch Changes
+
+- Fix validation logic
+
 ## 0.6.9
 
 ### Patch Changes