+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ballerine/workflows-service
 
+## 0.7.10
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.11
+  - @ballerine/workflow-node-sdk@0.6.11
+
+## 0.7.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.10
+  - @ballerine/workflow-node-sdk@0.6.10
+
 ## 0.7.8
 
 ### Patch Changes